world wip 2, variadic poc, chore

+++ ParaTH/src/Utils/SpanExtensions.cs
@@ -8,6 +8,6 @@ public static class SpanExtensions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static ref T UnsafeAt<T>(this Span<T> span, int index)
     {
-        return ref Unsafe.Add(ref MemoryMarshal.GetReference(span), index);
+        return ref Unsafe.Add(ref MemoryMarshal.GetReference(span), (nint)(uint)index);
     }
 }
